Changed file type for BulletPoint.
