+++ js/Series/Bullet/BulletPoint.js
@@ -1,0 +1,62 @@
+/* *
+ *
+ *  (c) 2010-2020 Torstein Honsi
+ *
+ *  License: www.highcharts.com/license
+ *
+ *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
+ *
+ * */
+var __extends = (this && this.__extends) || (function () {
+    var extendStatics = function (d, b) {
+        extendStatics = Object.setPrototypeOf ||
+            ({ __proto__: [] } instanceof Array && function (d, b) { d.__proto__ = b; }) ||
+            function (d, b) { for (var p in b) if (b.hasOwnProperty(p)) d[p] = b[p]; };
+        return extendStatics(d, b);
+    };
+    return function (d, b) {
+        extendStatics(d, b);
+        function __() { this.constructor = d; }
+        d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());
+    };
+})();
+import ColumnSeries from '../Column/ColumnSeries.js';
+/* *
+ *
+ *  Class
+ *
+ * */
+var BulletPoint = /** @class */ (function (_super) {
+    __extends(BulletPoint, _super);
+    function BulletPoint() {
+        var _this = _super !== null && _super.apply(this, arguments) || this;
+        _this.options = void 0;
+        _this.series = void 0;
+        return _this;
+        /* eslint-enable valid-jsdoc */
+    }
+    /* *
+     *
+     *  Functions
+     *
+     * */
+    /* eslint-disable valid-jsdoc */
+    /**
+     * Destroys target graphic.
+     * @private
+     */
+    BulletPoint.prototype.destroy = function () {
+        if (this.targetGraphic) {
+            this.targetGraphic = this.targetGraphic.destroy();
+        }
+        _super.prototype.destroy.apply(this, arguments);
+        return;
+    };
+    return BulletPoint;
+}(ColumnSeries.prototype.pointClass));
+/* *
+ *
+ *  Export Default
+ *
+ * */
+export default BulletPoint;
